Refactor options components
